+++ apps/nextjs-web/src/gql/gql.ts
@@ -15,13 +15,7 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
 const documents = {
   '\n      query GetLangAndLlm {\n        searchFilters {\n          language\n          largeLanguageModel\n          publicationState\n        }\n      }\n    ':
     types.GetLangAndLlmDocument,
-  '\n            query SummaryFeedbacksById($webPageSummaryId: String!) {\n              summaryFeedbacksById(\n                data: { webPageSummaryId: $webPageSummaryId }\n              ) {\n                id\n              }\n            }\n          ':
-    types.SummaryFeedbacksByIdDocument,
-  '\n              mutation deleteSummaryFeedbackMutation(\n                $summaryFeedbackId: String!\n              ) {\n                deleteSummaryFeedbackMutation(\n                  data: { summaryFeedbackId: $summaryFeedbackId }\n                ) {\n                  id\n                }\n              }\n            ':
-    types.DeleteSummaryFeedbackMutationDocument,
-  '\n              mutation updateSummaryFeedbackMutation(\n                $summaryFeedbackId: String!\n                $position: Int!\n                $voting: SummaryFeedbackVoting!\n                $webPageSummaryId: String!\n                $query: String!\n              ) {\n                updateSummaryFeedbackMutation(\n                  data: {\n                    summaryFeedbackId: $summaryFeedbackId\n                    position: $position\n                    voting: $voting\n                    webPageSummaryId: $webPageSummaryId\n                    query: $query\n                  }\n                ) {\n                  feedbackDate\n                  id\n                  position\n                  voting\n                  webPageSummaryId\n                  query\n                }\n              }\n            ':
-    types.UpdateSummaryFeedbackMutationDocument,
-  '\n              mutation createSummaryFeedback(\n                $position: Int!\n                $voting: SummaryFeedbackVoting!\n                $webPageSummaryId: String!\n                $query: String!\n              ) {\n                createSummaryFeedbackMutation(\n                  data: {\n                    position: $position\n                    voting: $voting\n                    webPageSummaryId: $webPageSummaryId\n                    query: $query\n                  }\n                ) {\n                  feedbackDate\n                  id\n                  position\n                  query\n                  voting\n                  webPageSummaryId\n                }\n              }\n            ':
+  '\n            mutation createSummaryFeedback(\n              $position: Int!\n              $voting: SummaryFeedbackVoting!\n              $webPageSummaryId: String!\n              $query: String!\n            ) {\n              createSummaryFeedback(\n                data: {\n                  position: $position\n                  voting: $voting\n                  webPageSummaryId: $webPageSummaryId\n                  query: $query\n                }\n              ) {\n                feedbackDate\n                id\n                position\n                query\n                voting\n                webPageSummaryId\n              }\n            }\n          ':
     types.CreateSummaryFeedbackDocument,
   '\n  fragment InfoCard on IndexedWebPage {\n    title\n    url\n    language\n    publicationState\n    keywords\n    summary\n  }\n':
     types.InfoCardFragmentDoc,
@@ -53,26 +47,8 @@ export function graphql(
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
-  source: '\n            query SummaryFeedbacksById($webPageSummaryId: String!) {\n              summaryFeedbacksById(\n                data: { webPageSummaryId: $webPageSummaryId }\n              ) {\n                id\n              }\n            }\n          ',
-): (typeof documents)['\n            query SummaryFeedbacksById($webPageSummaryId: String!) {\n              summaryFeedbacksById(\n                data: { webPageSummaryId: $webPageSummaryId }\n              ) {\n                id\n              }\n            }\n          ']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
-  source: '\n              mutation deleteSummaryFeedbackMutation(\n                $summaryFeedbackId: String!\n              ) {\n                deleteSummaryFeedbackMutation(\n                  data: { summaryFeedbackId: $summaryFeedbackId }\n                ) {\n                  id\n                }\n              }\n            ',
-): (typeof documents)['\n              mutation deleteSummaryFeedbackMutation(\n                $summaryFeedbackId: String!\n              ) {\n                deleteSummaryFeedbackMutation(\n                  data: { summaryFeedbackId: $summaryFeedbackId }\n                ) {\n                  id\n                }\n              }\n            ']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
-  source: '\n              mutation updateSummaryFeedbackMutation(\n                $summaryFeedbackId: String!\n                $position: Int!\n                $voting: SummaryFeedbackVoting!\n                $webPageSummaryId: String!\n                $query: String!\n              ) {\n                updateSummaryFeedbackMutation(\n                  data: {\n                    summaryFeedbackId: $summaryFeedbackId\n                    position: $position\n                    voting: $voting\n                    webPageSummaryId: $webPageSummaryId\n                    query: $query\n                  }\n                ) {\n                  feedbackDate\n                  id\n                  position\n                  voting\n                  webPageSummaryId\n                  query\n                }\n              }\n            ',
-): (typeof documents)['\n              mutation updateSummaryFeedbackMutation(\n                $summaryFeedbackId: String!\n                $position: Int!\n                $voting: SummaryFeedbackVoting!\n                $webPageSummaryId: String!\n                $query: String!\n              ) {\n                updateSummaryFeedbackMutation(\n                  data: {\n                    summaryFeedbackId: $summaryFeedbackId\n                    position: $position\n                    voting: $voting\n                    webPageSummaryId: $webPageSummaryId\n                    query: $query\n                  }\n                ) {\n                  feedbackDate\n                  id\n                  position\n                  voting\n                  webPageSummaryId\n                  query\n                }\n              }\n            ']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
-  source: '\n              mutation createSummaryFeedback(\n                $position: Int!\n                $voting: SummaryFeedbackVoting!\n                $webPageSummaryId: String!\n                $query: String!\n              ) {\n                createSummaryFeedbackMutation(\n                  data: {\n                    position: $position\n                    voting: $voting\n                    webPageSummaryId: $webPageSummaryId\n                    query: $query\n                  }\n                ) {\n                  feedbackDate\n                  id\n                  position\n                  query\n                  voting\n                  webPageSummaryId\n                }\n              }\n            ',
-): (typeof documents)['\n              mutation createSummaryFeedback(\n                $position: Int!\n                $voting: SummaryFeedbackVoting!\n                $webPageSummaryId: String!\n                $query: String!\n              ) {\n                createSummaryFeedbackMutation(\n                  data: {\n                    position: $position\n                    voting: $voting\n                    webPageSummaryId: $webPageSummaryId\n                    query: $query\n                  }\n                ) {\n                  feedbackDate\n                  id\n                  position\n                  query\n                  voting\n                  webPageSummaryId\n                }\n              }\n            ']
+  source: '\n            mutation createSummaryFeedback(\n              $position: Int!\n              $voting: SummaryFeedbackVoting!\n              $webPageSummaryId: String!\n              $query: String!\n            ) {\n              createSummaryFeedback(\n                data: {\n                  position: $position\n                  voting: $voting\n                  webPageSummaryId: $webPageSummaryId\n                  query: $query\n                }\n              ) {\n                feedbackDate\n                id\n                position\n                query\n                voting\n                webPageSummaryId\n              }\n            }\n          ',
+): (typeof documents)['\n            mutation createSummaryFeedback(\n              $position: Int!\n              $voting: SummaryFeedbackVoting!\n              $webPageSummaryId: String!\n              $query: String!\n            ) {\n              createSummaryFeedback(\n                data: {\n                  position: $position\n                  voting: $voting\n                  webPageSummaryId: $webPageSummaryId\n                  query: $query\n                }\n              ) {\n                feedbackDate\n                id\n                position\n                query\n                voting\n                webPageSummaryId\n              }\n            }\n          ']
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
