+++ packages/pothos-graphql/src/gql/gql.ts
@@ -15,18 +15,10 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
 const documents = {
   '\n            query GetUniqueValues {\n              webPageSummaries(publicationState: PREVIEW, locale: "all") {\n                data {\n                  attributes {\n                    locale\n                    largeLanguageModel\n                  }\n                }\n              }\n            }\n          ':
     types.GetUniqueValuesDocument,
-  '\n  fragment SummaryFeedback on SummaryFeedbackEntity {\n    id\n    attributes {\n      feedbackDate\n      position\n      query\n      voting\n      web_page_summary {\n        data {\n          id\n        }\n      }\n    }\n  }\n':
-    types.SummaryFeedbackFragmentDoc,
   '\n            mutation CreateSummaryFeedback($input: SummaryFeedbackInput!) {\n              createSummaryFeedback(data: $input) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ':
     types.CreateSummaryFeedbackDocument,
-  '\n            mutation DeleteSummaryFeedback($id: ID!) {\n              deleteSummaryFeedback(id: $id) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ':
-    types.DeleteSummaryFeedbackDocument,
-  '\n            mutation UpdateSummaryFeedback(\n              $id: ID!\n              $data: SummaryFeedbackInput!\n            ) {\n              updateSummaryFeedback(id: $id, data: $data) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ':
-    types.UpdateSummaryFeedbackDocument,
-  '\n            query getSummaryFeedbacks($webPageSummaryId: ID!) {\n              summaryFeedbacks(\n                filters: { web_page_summary: { id: { eq: $webPageSummaryId } } }\n              ) {\n                data {\n                  ...GetSummaryFeedback\n                }\n              }\n            }\n          ':
-    types.GetSummaryFeedbacksDocument,
-  '\n              fragment GetSummaryFeedback on SummaryFeedbackEntity {\n                id\n                attributes {\n                  web_page_summary {\n                    data {\n                      id\n                    }\n                  }\n                }\n              }\n            ':
-    types.GetSummaryFeedbackFragmentDoc,
+  '\n            fragment SummaryFeedback on SummaryFeedbackEntity {\n              id\n              attributes {\n                feedbackDate\n                position\n                query\n                voting\n                web_page_summary {\n                  data {\n                    id\n                  }\n                }\n              }\n            }\n          ':
+    types.SummaryFeedbackFragmentDoc,
   '\n            query GetWebPageSummaries {\n              webPageSummaries(publicationState: PREVIEW, locale: "all") {\n                data {\n                  ...WebPageSummary\n                }\n              }\n            }\n          ':
     types.GetWebPageSummariesDocument,
   '\n              fragment WebPageSummary on WebPageSummaryEntity {\n                id\n                attributes {\n                  locale\n                  keywords\n                  summary\n                  largeLanguageModel\n                  scraped_web_page {\n                    data {\n                      attributes {\n                        title\n                        url\n                        originalContent\n                        publishedAt\n                      }\n                    }\n                  }\n                }\n              }\n            ':
@@ -57,38 +49,14 @@ export function graphql(
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
-  source: '\n  fragment SummaryFeedback on SummaryFeedbackEntity {\n    id\n    attributes {\n      feedbackDate\n      position\n      query\n      voting\n      web_page_summary {\n        data {\n          id\n        }\n      }\n    }\n  }\n',
-): (typeof documents)['\n  fragment SummaryFeedback on SummaryFeedbackEntity {\n    id\n    attributes {\n      feedbackDate\n      position\n      query\n      voting\n      web_page_summary {\n        data {\n          id\n        }\n      }\n    }\n  }\n']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
   source: '\n            mutation CreateSummaryFeedback($input: SummaryFeedbackInput!) {\n              createSummaryFeedback(data: $input) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ',
 ): (typeof documents)['\n            mutation CreateSummaryFeedback($input: SummaryFeedbackInput!) {\n              createSummaryFeedback(data: $input) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ']
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
-  source: '\n            mutation DeleteSummaryFeedback($id: ID!) {\n              deleteSummaryFeedback(id: $id) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ',
-): (typeof documents)['\n            mutation DeleteSummaryFeedback($id: ID!) {\n              deleteSummaryFeedback(id: $id) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
-  source: '\n            mutation UpdateSummaryFeedback(\n              $id: ID!\n              $data: SummaryFeedbackInput!\n            ) {\n              updateSummaryFeedback(id: $id, data: $data) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ',
-): (typeof documents)['\n            mutation UpdateSummaryFeedback(\n              $id: ID!\n              $data: SummaryFeedbackInput!\n            ) {\n              updateSummaryFeedback(id: $id, data: $data) {\n                data {\n                  ...SummaryFeedback\n                }\n              }\n            }\n          ']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
-  source: '\n            query getSummaryFeedbacks($webPageSummaryId: ID!) {\n              summaryFeedbacks(\n                filters: { web_page_summary: { id: { eq: $webPageSummaryId } } }\n              ) {\n                data {\n                  ...GetSummaryFeedback\n                }\n              }\n            }\n          ',
-): (typeof documents)['\n            query getSummaryFeedbacks($webPageSummaryId: ID!) {\n              summaryFeedbacks(\n                filters: { web_page_summary: { id: { eq: $webPageSummaryId } } }\n              ) {\n                data {\n                  ...GetSummaryFeedback\n                }\n              }\n            }\n          ']
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(
-  source: '\n              fragment GetSummaryFeedback on SummaryFeedbackEntity {\n                id\n                attributes {\n                  web_page_summary {\n                    data {\n                      id\n                    }\n                  }\n                }\n              }\n            ',
-): (typeof documents)['\n              fragment GetSummaryFeedback on SummaryFeedbackEntity {\n                id\n                attributes {\n                  web_page_summary {\n                    data {\n                      id\n                    }\n                  }\n                }\n              }\n            ']
+  source: '\n            fragment SummaryFeedback on SummaryFeedbackEntity {\n              id\n              attributes {\n                feedbackDate\n                position\n                query\n                voting\n                web_page_summary {\n                  data {\n                    id\n                  }\n                }\n              }\n            }\n          ',
+): (typeof documents)['\n            fragment SummaryFeedback on SummaryFeedbackEntity {\n              id\n              attributes {\n                feedbackDate\n                position\n                query\n                voting\n                web_page_summary {\n                  data {\n                    id\n                  }\n                }\n              }\n            }\n          ']
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
