--- packages/pothos-graphql/src/summaryFeedback/mutations/index.ts
@@ -1,3 +1,0 @@
-import './summary-feedback-create-mutation-field'
-import './summary-feedback-delete-mutation-field#'
-import './summary-feedback-update-mutation-field'
