--- packages/pothos-graphql/src/summaryFeedback/queries/index.ts
@@ -1,1 +1,0 @@
-import './summary-feedback-query-field'
